configurate config.yml in configs
